✏️ [fix] 여행 참여 회원별 총 지출액 계산 API 요청 경로 변경
