+++ src/main/java/com/lets/go/right/now/domain/expense/controller/ExpenseController.java
@@ -76,4 +76,24 @@ public class ExpenseController {
             @RequestParam(value = "size", defaultValue = "4") int size) {
         return expenseService.getMyExpenses(tripId, customUserDetails.getEmail(), page, size);
     }
+
+    /**
+     * 여행 회원 별 총 지출액 반환 - 여행에 참여한 회원의 총 지출액 반환
+     */
+    @GetMapping("{trip_id}/member-expenses")
+    public ResponseEntity<?> getTravelMemberExpenses(
+            @PathVariable("trip_id") Long tripId) {
+        return expenseService.getTravelMemberExpenses(tripId);
+    }
+
+
+    /**
+     * 카테고리별 지출 리포트
+     */
+    @GetMapping("{trip_id}/category-report")
+    public ResponseEntity<?> getCategoryReport(
+            @PathVariable("trip_id") Long tripId) {
+        return expenseService.getCategoryReport(tripId);
+    }
+
 }
