+++ src/main/java/com/lets/go/right/now/domain/settlement/controller/SettlementController.java
@@ -67,13 +67,4 @@ public class SettlementController {
             @PathVariable("trip_id") Long tripId) {
         return settlementService.getTravelSettlementStatus(customUserDetails.getEmail(), tripId);
     }
-
-    /**
-     * 여행 회원 별 총 지출액 반환 - 여행에 참여한 회원의 총 지출액 반환
-     */
-    @GetMapping("{trip_id}/member-expenses")
-    public ResponseEntity<?> getTravelMemberExpenses(
-            @PathVariable("trip_id") Long tripId) {
-        return settlementService.getTravelMemberExpenses(tripId);
-    }
 }
